New Arrow arities: 1-0 and 1-3. Use Arrow10 as side effects to attach to sound-Arrows, for example to move the position around. Having them be sidecars to sound Arrows means they get sent the time data, so they have life breathed into them.
Arrow13 is an arrow that can actually compute a 3-d position changing in time. One instance we have is Rose, for making things fly around back and forth. This works in tandem with the Arrow10: The Arrow10 function calls an Arrow13 function and sets the position on a mixer, returning nothing.

Soup up the synthesizer's test method to emit a bunch of notes separately and together, so that I could listen hard for proof that the two pitches were flying around independently. I am 98% sure they are doing so.

From here I want to consolidate tech debt and offer a UI for waveforms, envelopes, Rose parameters and everything else.