Explore performance a bit. The AVAudioUnits add 10 CPU percentage points, so I commented out a couple. The AVAudioSourceNode's render block, even if returning a constant float, puts CPU at 33%. If I instead play an .aiff file and don't add the AVAudioSourceNode to the graph, CPU drops to 10%.
I moved all my Doubles to CoreFloat in case I could change CoreFloat to Float, but I can't because in fact the input from the audio engine (time) is in Double, and I want to be able to compose arrows, so doing (Double) -> Float would not work. The final output is Float but presumably the cast is not what's spiking the CPU.

This checkpoint can load .mid files through the file picker. There are several big bugs and one big change to be addressed: playing .mid files makes it clear I'm not playing all the notes, and there are pops as envelopes behave suddenly. CPU is bad, and could perhaps be mitigated by turning stuff off when not playing. And it's time to offer all the knobs in the UI, so I can iterate on the sound engine itself much more easily.